+++ packages/executor/package.json
@@ -33,6 +33,7 @@
   "dependencies": {
     "async-mutex": "0.4.0",
     "ethers": "5.7.2",
+    "@flashbots/ethers-provider-bundle": "0.6.2",
     "monitoring": "^1.0.24-alpha",
     "params": "^1.0.24-alpha",
     "types": "^1.0.24-alpha"
