+++ packages/types/src/executor/entities/MempoolEntry.ts
@@ -1,0 +1,8 @@
+export enum MempoolEntryStatus {
+  New = 0,
+  Pending = 1,
+  Submitted = 2,
+  IncludedToChain = 3,
+  Finalized = 4,
+  Cancelled = 5,
+}
